requirements.txt
aniso8601==9.0.1
attrs==22.2.0
click==8.1.3
exceptiongroup==1.1.0
Flask==2.2.3
flask-restx==1.0.6
Flask-SQLAlchemy==3.0.3
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.0
pluggy==1.0.0
pyrsistent==0.19.3
pytest==7.2.1
pytz==2022.7.1
smmap==5.0.0
SQLAlchemy==2.0.4
tomli==2.0.1
typing_extensions==4.5.0
Werkzeug==2.2.3
